requirements.txt
#Flask core
flask==2.3.2

# For requests
requests==2.31.0

# For name generation.
faker==19.2.0
# For environment variables
python-dotenv==1.0.0

# [code_style]-[BEGIN]
# Pre-commit hook.
pre-commit==3.3.3
# Upgrade syntax.
pyupgrade==3.10.1
# Formatter. Apply codestyle.
black==23.7.0
# Check codestyle.
flake8==6.1.0
# [code_style]-[END]
